backend/pipelines/svineflytning_pipeline/pyproject.toml
[project]
name = "svineflytning-pipeline"
version = "0.1.0"
description = "Bronze layer pipeline for fetching pig movement data from SvineflytningWS."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Martin Collignon", email = "your@email.com" },
]
dependencies = [
    "zeep~=4.3.1",
    "certifi~=2024.12.14",
    "google-cloud-secret-manager~=2.22.1",
    "lxml~=5.3.0",
    "xmlsec==1.3.14",
    "google-cloud-storage~=2.19.0",
    "python-dotenv~=1.0.1",
    "cryptography",
    "tqdm~=4.66.2",
    "pandas~=2.2.1",  # For data processing
    "xmltodict~=0.13.0",  # For handling raw XML responses
    "tenacity~=8.2.3",  # For retry logic
    "ijson~=3.2.3",  # For memory-efficient JSON streaming
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bronze", "silver"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0" 